Move Courses struct into courses.go, remove courses service.
