+++ term_department_courses.go
@@ -14,21 +14,6 @@ type TermDepartmentCoursesService struct {
 	client *Client
 }
 
-type Course struct {
-	Id                       int     `json:"id"`
-	Term                     string  `json:"term"`
-	CourseRegistrationNumber string  `json:"courseReferenceNumber"`
-	Subject                  string  `json:"subject"`
-	SectionNumber            string  `json:"sequenceNumber"`
-	Title                    string  `json:"courseTitle"`
-	IsOpen                   bool    `json:"openSection"`
-	CrossList                string  `json:"crossList"`
-	CrossListCapacity        int     `json:"crossListCapacity"`
-	CrossListAvailable       int     `json:"crossListAvailable"`
-	CreditHourHigh           float64 `json:"creditHourHigh"`
-	CreditHourLow            float64 `json:"creditHourLow"`
-}
-
 func (ts *TermDepartmentCoursesService) List(term, dept string) ([]Course, *http.Response, error) {
 	path := baseSearchPath
